Move app to lighty-examples, add placeholder READMEs
