+++ lighty-examples/lighty-bgp-community-restconf-app/README.md
@@ -1,0 +1,7 @@
+# Lighty BGP + RESTCONF Application
+Application starts the following components:
+* [Lighty Controller](../../lighty-core/lighty-controller/README.md)
+* [Lighty RESTCONF northbound plugin](../../lighty-modules/lighty-restconf-nb-community/README.md)
+* [Lighty BGP southbound plugin](../../lighty-modules/lighty-bgp/README.md)
+
+[comment]: <> (TODO, tutorial on how to connect BGP speaker, POSTMAN collection, topology export)
