+++ lighty-examples/lighty-bgp-community-restconf-app/pom.xml
@@ -11,9 +11,17 @@
         <relativePath>../../lighty-core/lighty-app-parent/pom.xml</relativePath>
     </parent>
 
-    <artifactId>lighty-bgp-app</artifactId>
+    <groupId>io.lighty.kit.examples.controllers</groupId>
+    <artifactId>lighty-bgp-community-restconf-app</artifactId>
     <version>16.0.0-SNAPSHOT</version>
     <packaging>jar</packaging>
+
+    <properties>
+        <application.main.class>io.lighty.examples.controllers.bgpapp.Main</application.main.class>
+        <application.attach.zip>true</application.attach.zip>
+
+        <maven.deploy.skip>true</maven.deploy.skip>
+    </properties>
 
     <dependencies>
         <dependency>
