+++ lighty-examples/lighty-bgp-community-restconf-app/src/main/java/io/lighty/examples/controllers/bgpapp/Main.java
@@ -5,7 +5,7 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at https://www.eclipse.org/legal/epl-v10.html
  */
-package io.lighty.applications.bgp.app;
+package io.lighty.examples.controllers.bgpapp;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Stopwatch;
