+++ lighty-examples/lighty-bgp-community-restconf-app/src/main/resources/log4j.properties
