+++ lighty-examples/pom.xml
@@ -24,5 +24,6 @@
         <module>lighty-community-restconf-netconf-app</module>
         <module>lighty-controller-springboot-netconf</module>
         <module>lighty-community-restconf-actions-app</module>
+        <module>lighty-bgp-community-restconf-app</module>
     </modules>
 </project>
