+++ lighty-modules/lighty-bgp/README.md
@@ -1,0 +1,1 @@
+[comment]: <> (TODO: Write once all bgpcep components are started)
